chore(atomic): merge fakeSummary controllers together
https://coveord.atlassian.net/browse/KIT-4431
